+++ src/functions/trayManager.js
@@ -127,6 +127,40 @@ function init(top, app, mainDir) {
                 },
             ]
         },
+        {
+            label: 'File',
+            submenu: [
+                isMac ? { role: 'close' } : { role: 'quit' }
+            ]
+        },
+        {
+            label: 'Edit',
+            submenu: [
+                { role: 'undo' },
+                { role: 'redo' },
+                { type: 'separator' },
+                { role: 'cut' },
+                { role: 'copy' },
+                { role: 'paste' },
+                ...(isMac ? [
+                    { role: 'pasteAndMatchStyle' },
+                    { role: 'delete' },
+                    { role: 'selectAll' },
+                    { type: 'separator' },
+                    {
+                        label: 'Speech',
+                        submenu: [
+                            { role: 'startspeaking' },
+                            { role: 'stopspeaking' }
+                        ]
+                    }
+                ] : [
+                    { role: 'delete' },
+                    { type: 'separator' },
+                    { role: 'selectAll' }
+                ])
+            ]
+        },
         // { role: 'viewMenu' }
         {
             label: 'View',
@@ -160,8 +194,6 @@ function init(top, app, mainDir) {
             ]
         }
     ];
-
-    console.log("AppMenuTemplate", appMenuTemplate)
 
     const appMenu = Menu.buildFromTemplate(appMenuTemplate);
     Menu.setApplicationMenu(appMenu);